+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog for [`hmep` package](http://hackage.haskell.org/package/hmep)
 
+## 0.1.0 *October 8th 2017*
+  * Breaking changes:
+    drop [monad-mersenne-random](http://hackage.haskell.org/package/monad-mersenne-random)
+    which doesn't build with newer version of GHC. Instead, depend on the
+    [probable](http://hackage.haskell.org/package/probable) package.
+
 ## 0.0.1 *October 7th 2017*
   * Improved demo: trigonometric identities solving example
   * Add `avgLoss` to the library